foo: fix nuttx build specifying invalid variable name
This also always installs on NuttX as well as on other systems. The
original condition was for demo application and that would make sense
but general approach is to have application you might want to compile in
NuttX and have it accessible. It might be also more beneficial to
disable the whole application as it is instead of just installation.