gitlab-ci: rework with gitlab CI cache and improve style
The CI now uses docker that provides various caching for Nix.

The new stage devenv was added to ensure that we build development
environment only once and reuse it from cache down the line.

nix develop is now quiet everywhere because it should be served from
cache and thus is no longer interesting.