libshvrpc: Add rpccall
This is pretty hacky but should simplify the method calling from SHVC.

The primary issue with SHV is that any message can be lost in the
network and thus we need to support multiple attempts and handling of
this on every call. At the same time we do want to support streaming and
thus handling of multiple call attempts must inevitably be part of the
application code.