Introduce the libshvclient with new idea for rpccall
This changes the previously introduced rpccall to be rather based on
simple callback. This is cleaner and more standard implementation in C.
In the end it also makes more sense because we can't react on not yet
verified message and thus we must buffer it anyway and act on it only
after that. We can of course do some preparations but nothing that
couldn't be reverted if it turns out that message at the end is being
aborted.