libshvrpc: fix and improve handler
This targets issues in threads synchronization in rpchandler.

This also implements support for sending signals from rpchandler thread.
In general this allows now multiple messages to be sent from handler
thread.

The activity keeping was moved from rpcclient to rpchandler. This is
primarilly because of the locking.

rpccall API is now documented.