docs: Replace RTD theme with Book theme
The reason for this is because Read The Docs theme merges page content
navigation with pages navigation. In the end it is way more beneficial
to have that split and that is what Book theme has.

The way the API is displayed suffers a bit with this switch but the
overall navigation in a bigger documents improves way more and thus this
switch is well-founded.