libshvrpc: rewrite logger to be lock free
The locking of shared logger introduces additional synchronization
between input and output that is not ensured every time. In general it
should not be an issue but problem is that with locks the program
behavior changes a log depending if logs are enabled or not.

This rather tries a different approach that uses line buffered output
and submits the whole lines for logged messages at once. The
disadvantage would in general be that we can't see not fully unpacked or
packed messages. The second disadvantage is the memory consumption
needed to buffer logs. The advantage is the faster and more consistent
behavior between multiple different logged clients.

This also ensures that helgrind and drd no longer detects error that was
caused by lock in logger. The issue there was that threads take lock for
both writing as well as reading but there are two different locks
associated with it in RPC handler and thus error and possible deadlock
was detected by helgrind and drd. Technically they were right but by
correctly working with API the deadlock would not be possible, but can
we rely on that?