libshvrpc: tweak for_cp_unpack_map implement Map support for dir
This provides map support for unpacking dir method results.

The for_cp_unpack_map is tweaked so the size provided must be size of
the longest string not the string to be recorded. Two additional bytes
are always added to recognize longer strings and to append null byte.
This is way more clear interface when used in the code.

This fixes few missing or invalid documentations.