libshvrpc: rework the rpcclient and stream layers
This is reworked version of rpcclient with improved and restructurated
stream layers. This original design that I pushed wasn't ideal in the
end and that is my fault of the judgement. The approach where all
streams are handled by the same code and in the end only minimal
functionality must be available to be hooked up for the TPC, Unix, Pipe
and TTY.

This also modifies the rpcclient itself to support reconnects and thus
be more in sync with abilities we will need in the future.