Add ability to disable TCP/IP RPC transport layer
TCP/IP is optional on NuttX and we want to use SHVC there without it.

This only disables implementation and causes initialization functions to
return NULL instead of objects. Thus existing code doesn't have to be
modified when TCP/IP is disabled; it can only fail if TCP/IP connection
is expected to be used.