libshvrpc: add support for RPC RI pattern matching
This implementation of pattern matching is basically fnmatch() function
but with additional double wildcard ** that is not supported on all
platforms. Therefore SHV specific implementation has to be used.

The pattern matching works for standard wildcard patterns '?', '*' and
'[' (including negation with '!') and follows POSIX.2, 3.13. It also
supports double wildcard '**' that matches multiple nodes.

Signed-off-by: Michal Lenc <michallenc@seznam.cz>