libshvrpc: Provide access to client from handler and fix obstack
This provides a way to access RPC client from handler. In general it is
not a good idea to access it because such access is not under handler's
protection but on the other hand for peer's name and disconnect it is
beneficial to have direct access to it this way without client
application having to keep its own reference alongside with reference to
the handler.

This also fixes access to the obstack from idle and message retrieval.