libshvrpc: rpcurl_parse now uses obstack instead of malloc
The previous implementation required a secret part of the structure to
hold duplicate pointers just to ensure that we can free malloc
allocated resources. By using obstack it is way more versatile while
providing way more easier free operation.

The primary reason for this change is configuration loading for the
shvcbroker that is being prepared.