libshvrpc/rpcclient_stream: cover case when read is called after eof
The serial stream implementation was using EOF on stream to prevent from
further message reading. The expectation was that feof must be cleared
with clearerr by user of the API thus read won't be called once EOF is
signaled unless we clear it. NuttX implementation thinks otherwise and
calls read even if EOF is set. The effect is that EOF might  be cleared
by fread, which is questionable. This was causing read timeout on
message flush.

As a side fix this fixes argument parsing for shvcp.