libshvrpc/rpchandler: Improve login handler
This ensure that applications running in the single loop are terminated
when login fails and that they can diagnose this. The demo application
now uses this feature.

The second change here is that idle function implementations can block
further ones in the stage sequence by returning negative number. This is
again used by login to suppress further stages. For msg function there
is no need to have anything special because that is how that stage works
(it is enough to return RPCHANDLER_MSG_DONE). On ls and dir it is not
desirable because all stages should disclose their nodes and methods;
also blocking is done already in msg and these are run only if msg did
not handle them.

The API change for the handler's message and idle handling was required
to provide a systematic way to request handler's loop termination and in
case of idle also to allow further stages suppression.