shvcs: Implementation
This is the first implementation of shvcs tool. It is intended as a way
to simply receive signals for some subscriptions.

This also includes dedicated handler for signals that can be used
outside of the shvcs.

The rename from original shvcsub is only to make name shorter and more
ambiguous; now it can be not only "subscribe" but also "signal".