libshvcallee,libshvcaller: merge to libshvrpc
The original idea was to provide utilities that are required for caller
and called applications separatelly to ease the integration but it turns
out that it includes just two files in both cases. It seems that
additional extensions for these libraries is not going to be significant
and thus it is way more easier to just have that merged as part of
libshvrpc. The effect is that everyone will link with libshvrpc directly
(unless you also need broker functionality).