libshvrpc/rpcclient_stream: prevent from deadlock on message sending
It is possible that connection stays active and valid (especially with
tty) but peer is no longer receiving data. This will result into buffers
to be filled to the point where not even one byte can be written. We
want to ensure that we can communicate even over slower links but if
link is stuck then we need to detect that and stop sending data to it to
prevent block.

This introduces kind of hacky way with rpcclient's errnum to record two
situations:
* When the current message can't be sent and thus some bytes were
  skipped. In such case it is not worth it to continue sending any data.
* When the previous message wasn't delivered due to full buffer. In
  this case we want to try to send some data just to see if situation
  did not improve in the meantime, but it is not desirable to was
  because that would force us to choose too small timeout for write
  (because otherwise our throughput would be very bad). Instead in this
  situation it only checks if buffer is full or not without any timeout.