libshvcp: fix warning when cpitem_extract_uint is used with uint64_t
The previous implementation was valid because outer condition would
always evaluate to false. But compiler still emitted warning although
the code in the true branch would never be executed.

This just changes check to always compare value but use less or equal
and compare against maximum value instead of the one over maximum.