libshvrpc/handler: remove rpchandler_msg_access_level
This promotes and invalid pattern to check access level only after
argument parsing while it is way better to check it before we even
attempt to parse arguments. In other words: the check for access level
should be part of the check if path and method is even something we want
to check.

This also provides utility for sending signals from RPC handler.