libshvcp: remove RAW operations
The original idea for RAW was to bypass unpacking and packing and write
potentially from one unpacker to packer without interpreting data. The
primary use case was for RPC Broker but in the end due to logging it is
better to always interpret data. It also provides only minimal overhead.
Now it introduces only case that must be explicitly handled in multiple
branches and has no real use case. It is better to just remove this
functionality.

The generic packers and unpackers that were introduced in this library
also made this more complicated because it is unclear if bytes would
even be compatible if they would not be processed by all functions in
them.