libshvrpc: add ability to query if client has connection tracking
We need this to optimize processes such as disconnect or waiting for
login in the broker. The presence of the connection tracking is
important because if there is no connection tracking then disconnecting
the client won't yield the expected behavior as well as client destroy
will most likely lead to only its recreation immediately.