libshvbroker: use rpcclient_contrack instead of parameter
This changes how we deduce if client should have infinite timeout for
login. The previous implementation was using parameter and dedicated
value for role that was actually an invalid pointer. Now we use
rpcclient_contrack instead. The idea is that if connection doesn't
support connection tracking then it is not worth it to rest it because
server will create a new client immediately because it has no way to
propagate our disconnect decision to the other side.