libshvhistory: introduce common history handler for records and files
This merges two separate handlers - records and files - to one common
history handler provided in rpchandler_history.c file. This is an API
breaking change as two existing handlers are removed and new one is
introduced. The rpchistory.h API (packing and unpacking) is not changed.

There are two major reasons for this change. One is a simpler addition of
new logging facilities: now just an array passed as an argument, no need
to have a handler per facility. Also the common handler will provide a
better starting ground for getLog method implementation.

Signed-off-by: Michal Lenc <michallenc@seznam.cz>