libshvhistory: implement getLog method to RPC History Handler
This commit adds getLog method handling to RPC History Handler. An array
of paths to recorded signals have to be passed as an argument to
rpchandler_history_new function. These paths are used to create a
virtual tree in .history.

The getLog method is implemented via user defined callback that packs
the correct records based on parameters given in the method call.

Signed-off-by: Michal Lenc <michallenc@seznam.cz>