libshvcp: extract to integer regardless the destination signedness
It is beneficial to be able to not allow only extraction of signed
integers to the signed types and unsigned respectively. Sometimes it
makes a lot of sense to send value as integer for example but have its
value in the unsigned variable. Extraction now correctly covers these
uses. Previously mixing unsigned and signed types would result into an
possible extraction validity but invalid value.