libshvcp: fix the unpack of big negative integers and date and time
There are two issues fixed in this commit.

The primary one is that big negative integers were invalidly masked
because mask was deduced only fron long and not from uintmax_t.

The secondary issue is that date and time must use division and not
shifting to the right because date and time is numerical signed integer
and thus it is aritmetic instead of logical shift.