libshvrpc: fix memory usage after free with rpchandler_ls_result_vfmt
This was caused by invalid assumption that on failure to add item the
caller will free the string passed to the shv_strset_add_dyn on its own.
That was actually done in rpchandler_ls_result_vfmt unconditionally and
thus freeing the used string. This instead moves the memory ownership
fully to the strset now and thus removes this bug and inconsistency.