libshvrpc: fix pipe usage causing deadlock
The pipe has a specific handling defined. The writes to the pipe are
always blocking (unless the buffer to be written is greater than system
buffer). Thus pipes must be set to non-blocking if we want to ensure
that we won't get stuck.

The issue is that we rely on write to do short write and return as soon
as no more bytes can be written. We return back to polling in such case
to check for timeout. We have no timeout setup for write and thus filled
pipe could halt the broker execution with no timeout. This is of course
not acceptable because one unresponsive process could halt the broker's
execution infinitely.