libshvrpc: hotfix unix sockets on NuttX
This is only a hotfix of the invalid NuttX implementation of the unix
sockets. Unix sockets should in general be like internet sockets but
NuttX implements them with pipes and thus write and read operations have
different nuances. The important difference for us is that writes to
pipes are not partial unless buffer is greater than the system one. This
causes write to block and because we have no chance to know how many
bytes we can write before this happens the FD must be switched to
non-blocking mode to prevent from this block happening.

The block described would happen in the rpcclient_stream and we want to
have that implementation independent on the blocking/nonblocking setting
(which is possible with exception of the pipes).