libshvcp: add cpdttots and cptstodt functions
There are functions to work with `struct tm` but lot of the time
timezone is not that important and thus `struct timespec` is enough and
this provides direct conversion from that data type to CP date time.

It is implemented as inlined function because it is really just a simple
calculation, but by including it in the library we will simplify the
code working with CP date time format.