libshvrpc: remove no longer valid RPC errors
SHV standard removed some invalid errors that were not actually used in
the SHV communication. Some of them are now marked as obsolete and thus
were removed from SHVC and some of them are replaced by implementation
specific definition. One of these errors is allocated by SHVC to encode
unknown error and other two are kept to the user applications for use.