docs: switch from doxygen to hawkmoth
Doxygen has a lot of issues when reading more advanced C code. It
doesn't fully understand macros. It doesn't support unnamed structures
and unions types. In combination with breathe it has issue with linking
documented items to each other. And other issues like this...

In total it is hard to keep doxygen documented API documented correctly.
The advantage of hawkmoth is that it uses clang directly and thus it can
read any code that clang can compile (which is something we expected
in our C projects). Thus descibed issues should no longer be relevant.
Of course there can be other issues but we must try hawkmoth out to
discover them.