shvc: remove long arguments and allow specific User ID
The application using shvc might want to provide custom start to the
user ID thus it should be possible to be provided as argument.

The long option --userid is replaced with short -s because I don't think
that introducing long options to the shvc is worth it at the moment.