rpclogger: avoid packing to CPON if logging doesn't make sense
Applications may decide whether logging messages are displayed or not
(with setlogmask if syslog logging is used for example). If logging
messages are not displayed, they don't have to be packed in CPON, which
is a performance bottleneck.

This adds callback function rpclogger_allowed_t that asks if logging
is allowed and returns the information to the caller. The logging
proceeds only if true is returned, otherwise packing is skipped. Passing
NULL to the rpclogger_new means the logging is always performed.

The first argument to rpclogger_new is now a pointer to the structure
rpclogger_callbacks unifying callbacks.

Signed-off-by: Michal Lenc <michallenc@seznam.cz>