libshvrpc: initialize obstack for every rpchandler_next
We suspect that obstack sometimes gets corrupted and thus initializing
it every time should remove sporadic issues we are encountering.

This is rather just a test of hypothesis and not fix of the real
problem. It is beneficial to reuse obstack but if we are corrupting it
somehow then it makes sense for us to reinitialize it every time to
check if that solves that issue.