libshvrpc: ensure correct handling of message validation
This primarilly adds attribute `warn_unused_result` to the
`rpchandler_msg_valid` to ensure that caller won't ignore it. It must be
called but nothing should be sent if `false` is returned. Thus the
result can't be ignored.

We could solve this by adding runtime check as well (to not allow
packing message unless validation is called but that would provide
`NULL` from `rpchandler_msg_new_response` and thus null pointer
dereference would be caused. That is harder to track than this warning.
Thus we still allow sending messages without prior validation in the
code itself.