docs: add hawkmoth_clang meson source template
This adds template that is commonly required if dependency is specified
in Meson build system and to get it to the clang we need to get include
paths. This code uses Meson's introspection to get this information.

This commit also removes path insert as that is not required because we
are not documenting Python project and thus there are no Python modules
in the project's root we need to access.

To make it actually work this also fixes requirement for valgrind and
makes it truly optional.