treewide: suppress clang-tidy false-positives
These are new or old false positives reported by clang-tidy.

In case of strset it is clear that clang-tidy fails to understand the
code and follows two assumptions that are exclusive with each other.
Thus reporting impossible branch.

In case of va_end clang has bug and is unable to link va_start that is
right when it starts to analyze the code or even right in the range of
analyze but for some reason ignore (possibly because it is part of the
same macro and thus on the same line).