libshvrpc: fix possible SIGHUP on message sending to stream
There is a possible condition where we do not know yet about the peer
disconnect. In such a case when we are writing a message to it we could
encounter SIGHUP halting the application.

The delivery of SIGHUP happens if read or write operation is used again
after HUP was already reported. It seems that in this case poll is
considered the first HUP delivery and thus we can't ignore it unless we
want to get SIGHUP signal.