libshvbroker: fix support for request aborts and response delays
This finishes changes done in c6cfc1bf300f1e946136526d95fb3de44a708fe1.
That commit provided basic support for request aborts and response
delays but broker's support was added as TODO. Actually, the support for
the broker is mostly just propagation to the mounted clients and nothing
more so these changes are enough to get support for them in the broker,
considering that handler doesn't support delayed responses yet.