libshvrpc: fix missing acknowledge on CAN
If CAN transport didn't receive or missed the acknowledge then it would
get stuck in the loop waiting for it and never sending any more frames.
The result was stuck connection that could have been recovered only with
disconnect frame.

The fix consist of improved behavior where we send the first frame again
(up to five times for more than five seconds timeout). If no acknowledge
is received then we just ignore it and continue.